Fix spotbugs issue

+++ java/system-test/system-test-drivers/src/main/java/sleeper/systemtest/drivers/query/MultipleQueries.java
@@ -136,7 +136,7 @@ public class MultipleQueries {
             }
         }
         LoggedDuration duration = LoggedDuration.between(startTime, Instant.now());
-        double rate = totalResults / duration.getSeconds();
+        double rate = totalResults / (double) duration.getSeconds();
         LOGGER.info("{} records returned in {} seconds at {} per second)", totalResults, duration, String.format("%.2f", rate));
     }
